Claude/fix catbench model loading l7tp p (#55)
* Refactor catbench to use exllamav3's model_init instead of custom loading

Replace the manual model loading in catbench.py (Config.from_directory →
Model.from_config → model.load → Cache) with exllamav3's standard
model_init.init(), matching how chat.py and all other exllamav3 examples
handle model loading. This fixes multi-GPU support and inherits automatic
device mapping, cache sizing, and future model_init improvements.

- catbench.py: Use model_init.add_args() for CLI and model_init.init()
  for loading; remove custom device parsing and manual model construction
- repo.py: Pass -cs (cache size) and -gs (gpu split) instead of -d;
  set CUDA_VISIBLE_DEVICES in subprocess env for device control


* Add token progress and diagnostic output to catbench inference loop

The inference loop was completely silent between CATBENCH_MODEL_LOADED
and CATBENCH_SAMPLE_DONE, making it look frozen on large models. Now
prints cache/stop condition info after load, sample start markers,
token count every 100 tokens with TPS, and final per-sample stats
including stop reason.


* Add VRAM cap, token progress display, and unified multi-GPU progress

Three catbench improvements:

1. VRAM filter: Models too large even for all GPUs combined are now
   skipped with a warning instead of being queued and failing.

2. Token progress: catbench.py emits CATBENCH_TOKENS markers every 100
   tokens during inference. _run_catbench_subprocess renders these as
   token count + TPS in the GPU status line, replacing the old 80/20
   progress bar model with: 0-100% bar during load, then live token
   count during inference.

3. Unified multi-GPU progress: Multi-GPU catbench jobs now use the same
   progress bar system as single-GPU (threaded subprocess + results
   queue + ANSI GPU status lines) instead of raw stdout printing.


* Strip think tags and code fences from response before SVG extraction

Reasoning models wrap output in <think> tags and code fences that can
interfere with SVG extraction. Clean the response text before passing
to extract_svg.


* Smooth loading progress bar by tracking model_init output lines

Instead of jumping to a fixed 50% during loading, start at 0% and
increment 3% per stdout line from model_init (capped at 95%), then
jump to 100% on CATBENCH_MODEL_LOADED. Throttled to update every 0.3s.


* Rename fp16 to bf16 and checkpoint on .txt files too

- Rename "fp16" to "bf16" everywhere: file prefixes, README grid
  labels, canonical filenames, and tests. Weights are typically bf16.

- Catbench checkpointing now also considers .txt sample files (failed
  SVG extractions) as evidence the model already ran, via the new
  _catbench_has_output() helper. Previously only canonical .svg was
  checked, causing re-runs after extraction failures.


---------

+++ ezexl3/catbench.py
@@ -170,7 +170,7 @@ def run_catbench(args) -> list:
 
     # Determine file naming
     if label == "bf16" or label == "base":
-        file_prefix = "fp16"
+        file_prefix = "bf16"
     else:
         try:
             val = float(label)
